+++ src/main/java/com/teamchallenge/bookti/controller/RegistrationController.java
@@ -1,0 +1,58 @@
+package com.teamchallenge.bookti.controller;
+
+import com.teamchallenge.bookti.dto.registration.NewUserRegistrationRequest;
+import com.teamchallenge.bookti.dto.registration.RegistrationResponse;
+import com.teamchallenge.bookti.dto.ErrorResponse;
+import com.teamchallenge.bookti.service.UserService;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
+
+@RequiredArgsConstructor
+@RequestMapping("/registration")
+@RestController
+public class RegistrationController {
+
+    private final UserService defaultUserService;
+
+    @Operation(
+            summary = "User registration",
+            responses = {
+                    @ApiResponse(
+                            responseCode = "201",
+                            description = "User created successfully",
+                            content = {
+                                    @Content(
+                                            mediaType = APPLICATION_JSON_VALUE,
+                                            schema = @Schema(implementation = RegistrationResponse.class)
+                                    )
+                            }
+                    ),
+                    @ApiResponse(
+                            responseCode = "400",
+                            description = "Bad request",
+                            content = {
+                                    @Content(
+                                            mediaType = APPLICATION_JSON_VALUE,
+                                            schema = @Schema(implementation = ErrorResponse.class))
+                            }
+                    )
+            }
+    )
+    @PostMapping(path = "/register", consumes = {APPLICATION_JSON_VALUE}, produces = {APPLICATION_JSON_VALUE})
+    public ResponseEntity<RegistrationResponse> register(@Valid @RequestBody NewUserRegistrationRequest newUserRegistrationRequest) {
+        return ResponseEntity.status(HttpStatus.CREATED).body(defaultUserService.create(newUserRegistrationRequest));
+    }
+}
+
