--- src/main/java/com/teamchallenge/bookti/registration/RegistrationController.java
@@ -1,42 +1,0 @@
-package com.teamchallenge.bookti.registration;
-
-import com.teamchallenge.bookti.exception.ErrorResponse;
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.media.Schema;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import io.swagger.v3.oas.annotations.responses.ApiResponses;
-import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
-
-@RequiredArgsConstructor
-@RequestMapping("/registration")
-@RestController
-public class RegistrationController {
-    private final UserService userService;
-
-    @Operation(summary = "User registration")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "201",
-                    description = "User created successfully",
-                    content = {@Content(mediaType = APPLICATION_JSON_VALUE,
-                            schema = @Schema(implementation = RegistrationResponse.class))}),
-            @ApiResponse(responseCode = "400",
-                    description = "Bad request",
-                    content = {@Content(mediaType = APPLICATION_JSON_VALUE,
-                            schema = @Schema(implementation = ErrorResponse.class))})
-    })
-    @PostMapping("/register")
-    public ResponseEntity<RegistrationResponse> register(@Valid @RequestBody UserDto userDto) {
-        return ResponseEntity.status(HttpStatus.CREATED).body(userService.createUser(userDto));
-    }
-}
-
